requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.2
djangorestframework==3.16.0
frozenlist==1.6.0
idna==3.10
modeltranslation==0.25
multidict==6.4.3
openai==0.28.0
pillow==11.2.1
propcache==0.3.1
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.0
